Added custom hail string escape / unescape functions.  All inputs should be escaped (parse error if not)
