+++ src/main/scala/org/broadinstitute/hail/io/annotators/GlobalTableAnnotator.scala
@@ -5,6 +5,7 @@ import org.apache.spark.sql.Row
 import org.broadinstitute.hail.Utils._
 import org.broadinstitute.hail.annotations._
 import org.broadinstitute.hail.expr._
+import org.broadinstitute.hail.utils.StringEscapeUtils
 
 import scala.collection.mutable
 
@@ -15,7 +16,7 @@ object GlobalTableAnnotator extends TSVAnnotator {
       if (lines.isEmpty)
         fatal("empty file")
 
-      val delimiter = unescapeString(delim)
+      val delimiter = StringEscapeUtils.unescapeString(delim)
 
       val header = lines.next().value
       val split = header.split(delimiter)
