+++ src/main/scala/org/broadinstitute/hail/io/annotators/SampleTableAnnotator.scala
@@ -5,6 +5,7 @@ import org.apache.spark.sql.Row
 import org.broadinstitute.hail.Utils._
 import org.broadinstitute.hail.annotations._
 import org.broadinstitute.hail.expr._
+import org.broadinstitute.hail.utils.StringEscapeUtils._
 
 import scala.collection.mutable
