+++ src/main/scala/org/broadinstitute/hail/io/annotators/VariantTableAnnotator.scala
@@ -4,6 +4,7 @@ import org.apache.spark.SparkContext
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.Row
 import org.broadinstitute.hail.Utils._
+import org.broadinstitute.hail.utils.StringEscapeUtils._
 import org.broadinstitute.hail.annotations.Annotation
 import org.broadinstitute.hail.expr._
 import org.broadinstitute.hail.variant.Variant
