+++ src/main/scala/org/broadinstitute/hail/utils/StringEscapeUtils.scala
@@ -1,0 +1,317 @@
+package org.broadinstitute.hail.utils
+
+import java.util.Locale
+
+import org.broadinstitute.hail.Utils._
+
+object StringEscapeUtils {
+
+  def hex(ch: Char): String = Integer.toHexString(ch).toUpperCase(Locale.ENGLISH)
+
+  def escapeUnprintable(str: String): String = escapeUnprintable(str, new StringBuilder(capacity = str.length * 2))
+
+  def escapeUnprintable(str: String, sb: StringBuilder): String = {
+    sb.clear()
+
+    var sz: Int = 0
+    sz = str.length
+    var i: Int = 0
+    while (i < sz) {
+      {
+        val ch: Char = str.charAt(i)
+        if (ch > 0xfff) {
+          sb.append("\\u" + hex(ch))
+        }
+        else if (ch > 0xff) {
+          sb.append("\\u0" + hex(ch))
+        }
+        else if (ch > 0x7f) {
+          sb.append("\\u00" + hex(ch))
+        }
+        else if (ch < 32) {
+          ch match {
+            case '\b' =>
+              sb += '\\'
+              sb += 'b'
+            case '\n' =>
+              sb += '\\'
+              sb += 'n'
+            case '\t' =>
+              sb += '\\'
+              sb += 't'
+            case '\f' =>
+              sb += '\\'
+              sb += 'f'
+            case '\r' =>
+              sb += '\\'
+              sb += 'r'
+            case _ =>
+              if (ch > 0xf) {
+                sb.append("\\u00" + hex(ch))
+              }
+              else {
+                sb.append("\\u000" + hex(ch))
+              }
+          }
+        }
+        else {
+          sb.append(ch)
+        }
+      }
+      i += 1
+    }
+    sb.result()
+  }
+
+  def escapeString(str: String): String = escapeString(str, new StringBuilder(capacity = str.length * 2))
+
+  def escapeString(str: String, sb: StringBuilder): String = {
+    sb.clear()
+
+    var sz: Int = 0
+    sz = str.length
+    var i: Int = 0
+    while (i < sz) {
+      {
+        val ch: Char = str.charAt(i)
+        if (ch > 0xfff) {
+          sb.append("\\u" + hex(ch))
+        }
+        else if (ch > 0xff) {
+          sb.append("\\u0" + hex(ch))
+        }
+        else if (ch > 0x7f) {
+          sb.append("\\u00" + hex(ch))
+        }
+        else if (ch < 32) {
+          ch match {
+            case '\b' =>
+              sb += '\\'
+              sb += 'b'
+            case '\n' =>
+              sb += '\\'
+              sb += 'n'
+            case '\t' =>
+              sb += '\\'
+              sb += 't'
+            case '\f' =>
+              sb += '\\'
+              sb += 'f'
+            case '\r' =>
+              sb += '\\'
+              sb += 'r'
+            case _ =>
+              if (ch > 0xf) {
+                sb.append("\\u00" + hex(ch))
+              }
+              else {
+                sb.append("\\u000" + hex(ch))
+              }
+          }
+        }
+        else {
+          ch match {
+            case '\'' =>
+              sb += '\\'
+              sb += '\''
+            case '\"' =>
+              sb += '\\'
+              sb += '\"'
+            case '\\' =>
+              sb += '\\'
+              sb += '\\'
+            case _ =>
+              sb.append(ch)
+          }
+        }
+      }
+      i += 1
+    }
+    sb.result()
+  }
+
+  def escapeBackticked(str: String): String = escapeBackticked(str, new StringBuilder(capacity = str.length * 2))
+
+  def escapeBackticked(str: String, sb: StringBuilder): String = {
+    sb.clear()
+
+    var sz: Int = 0
+    sz = str.length
+    var i: Int = 0
+    while (i < sz) {
+      {
+        val ch: Char = str.charAt(i)
+        if (ch > 0xfff) {
+          sb.append("\\u" + hex(ch))
+        }
+        else if (ch > 0xff) {
+          sb.append("\\u0" + hex(ch))
+        }
+        else if (ch > 0x7f) {
+          sb.append("\\u00" + hex(ch))
+        }
+        else if (ch < 32) {
+          ch match {
+            case '\b' =>
+              sb += '\\'
+              sb += 'b'
+            case '\n' =>
+              sb += '\\'
+              sb += 'n'
+            case '\t' =>
+              sb += '\\'
+              sb += 't'
+            case '\f' =>
+              sb += '\\'
+              sb += 'f'
+            case '\r' =>
+              sb += '\\'
+              sb += 'r'
+            case _ =>
+              if (ch > 0xf) {
+                sb.append("\\u00" + hex(ch))
+              }
+              else {
+                sb.append("\\u000" + hex(ch))
+              }
+          }
+        }
+        else {
+          ch match {
+            case '\\' =>
+              sb += '\\'
+              sb += '\\'
+            case '`' =>
+              sb += '\\'
+              sb += '`'
+            case _ =>
+              sb.append(ch)
+          }
+        }
+      }
+      i += 1
+    }
+    sb.result()
+  }
+
+  def unescapeString(str: String): String = unescapeString(str, new StringBuilder(capacity = str.length))
+
+  def unescapeString(str: String, sb: StringBuilder): String = {
+    sb.clear()
+
+    val sz = str.length()
+    var hadSlash = false
+    var inUnicode = false
+    lazy val unicode = new StringBuilder(capacity = 4)
+    var i = 0
+    while (i < str.length) {
+
+      val ch = str.charAt(i)
+      if (inUnicode) {
+        // if in unicode, then we're reading unicode
+        // values in somehow
+        unicode.append(ch)
+        if (unicode.length == 4) {
+          // unicode now contains the four hex digits
+          // which represents our unicode character
+          try {
+            val value = Integer.parseInt(unicode.toString(), 16)
+            sb += value.toChar
+            unicode.clear()
+            inUnicode = false
+            hadSlash = false
+          } catch {
+            case nfe: NumberFormatException =>
+              fatal("Unable to parse unicode value: " + unicode)
+          }
+        }
+      }
+      else if (hadSlash) {
+        hadSlash = false
+        ch match {
+          case '\\' => sb += '\\'
+          case '\'' => sb += '\''
+          case '\"' => sb += '\"'
+          case 'r' => sb += '\r'
+          case 'f' => sb += '\f'
+          case 't' => sb += '\t'
+          case 'n' => sb += '\n'
+          case 'b' => sb += '\b'
+          case 'u' => inUnicode = true
+          case _ => fatal(s"Got invalid string escape character: `\\$ch'")
+        }
+      }
+      else if (ch == '\\')
+        hadSlash = true
+      else
+        sb += ch
+      i += 1
+    }
+    if (hadSlash) {
+      // then we're in the weird case of a \ at the end of the
+      // string, let's output it anyway.
+      sb += '\\'
+    }
+    sb.result()
+  }
+
+  def unescapeBackticked(str: String): String = unescapeBackticked(str, new StringBuilder(capacity = str.length))
+
+  def unescapeBackticked(str: String, sb: StringBuilder): String = {
+    sb.clear()
+
+    val sz = str.length()
+    var hadSlash = false
+    var inUnicode = false
+    lazy val unicode = new StringBuilder(capacity = 4)
+    var i = 0
+    while (i < str.length) {
+
+      val ch = str.charAt(i)
+      if (inUnicode) {
+        // if in unicode, then we're reading unicode
+        // values in somehow
+        unicode.append(ch)
+        if (unicode.length == 4) {
+          // unicode now contains the four hex digits
+          // which represents our unicode character
+          try {
+            val value = Integer.parseInt(unicode.toString(), 16)
+            sb += value.toChar
+            unicode.clear()
+            inUnicode = false
+            hadSlash = false
+          } catch {
+            case nfe: NumberFormatException =>
+              fatal("Unable to parse unicode value: " + unicode)
+          }
+        }
+      }
+      else if (hadSlash) {
+        hadSlash = false
+        ch match {
+          case '\\' => sb += '\\'
+          case '`' => sb += '`'
+          case 'r' => sb += '\r'
+          case 'f' => sb += '\f'
+          case 't' => sb += '\t'
+          case 'n' => sb += '\n'
+          case 'b' => sb += '\b'
+          case 'u' => inUnicode = true
+          case _ => fatal(s"Got invalid backticked escape character: `\\$ch'")
+        }
+      }
+      else if (ch == '\\')
+        hadSlash = true
+      else
+        sb += ch
+      i += 1
+    }
+    if (hadSlash) {
+      // then we're in the weird case of a \ at the end of the
+      // string, let's output it anyway.
+      sb += '\\'
+    }
+    sb.result()
+  }
+}
