+++ src/test/scala/org/broadinstitute/hail/methods/FilterSuite.scala
@@ -1,9 +1,9 @@
 package org.broadinstitute.hail.methods
 
-import org.broadinstitute.hail.SparkSuite
 import org.broadinstitute.hail.driver._
 import org.broadinstitute.hail.expr._
 import org.broadinstitute.hail.utils.TestRDDBuilder
+import org.broadinstitute.hail.{FatalException, SparkSuite}
 import org.testng.annotations.Test
 
 class FilterSuite extends SparkSuite {
@@ -170,6 +170,9 @@ class FilterSuite extends SparkSuite {
     val state = SplitMulti.run(State(sc, sqlContext, vds), Array.empty[String])
     val s2 = FilterVariantsExpr.run(state, Array("--keep", "-c", "va.`weird name \\t test` > 500"))
     assert(s2.vds.nVariants == vds.nVariants)
+
+    val e = intercept[FatalException](FilterVariantsExpr.run(state, Array("--keep", "-c", "va.`bad\tinput` == 5")))
+    assert(e.getMessage.contains("invalid character in backtick identifier: `\\t'"))
   }
 
   @Test def testPAB() {
